+++ src/curies/mapping_service/rdflib_custom.py
